refactor(json-assertions): split JsonAssert responsibilities into static delegates
Why:
- JsonAssert had mixed responsibilities for path resolution, value checks, schema validation, and message formatting.

What Changed:
- Added JsonAssertionContext and extracted JsonPathResolver, JsonValueAssertion, JsonSchemaValidator, and JsonSchemaValidationMessageBuilder.
- Reduced JsonAssertion to a facade that delegates to static helper modules.
- Preserved existing fluent API signatures and failure contracts.

Impact:
- No production behavior impact because changes are limited to test helper internals.
- Existing JsonAssert call sites remain compatible.
